+++ web/CHANGES.md
@@ -1,6 +1,7 @@
 # unreleased 0.6.3
 ## Add
 - add default impl to `handler::state::BorrowState` trait for `Box`, `Rc` and `Arc` types
+- add `middleware::WebContext`
 
 ## Change
 - revert `handler::state::BorrowState` change from `0.6.2`